backend/requirements.txt
asgiref
Django
django-cors-headers
django-otp
django-braces
djangorestframework
django-extensions
djangorestframework-simplejwt
psycopg2-binary
django-quill-editor
PyJWT
PyMySQL
pymysql
django-storages
python-dotenv
arrow
django-environ
pytz
sqlparse
tzdata
pytest
gunicorn
pytest-cov
pytest-django
cryptography
daphne
channels
channels_yroom
django-jazzmin
Pillow
django_quill
redis
channels_redis
django-ckeditor
dotenv
environ
jazzmin
yroom
